issue: basic curd

+++ src/main/java/cn/jzyunqi/ms/article/IArticleDao.java
@@ -2,5 +2,5 @@ package cn.jzyunqi.ms.article;
 
 import cn.jzyunqi.common.persistence.dao.BaseDao;
 
-public interface IArticleDao extends BaseDao<Article, Integer> {
+public interface IArticleDao extends BaseDao<Article, Long> {
 }
